Handles CMake policy CMP0063
The ``OLD`` behavior for this policy is to ignore the visibility properties
for static libraries, object libraries, and executables without exports.
The ``NEW`` behavior for this policy is to honor the visibility properties
for all target types.

This policy was introduced in CMake version 3.3.  CMake version
3.3.0 warns when the policy is not set and uses ``OLD`` behavior. Use
the ``cmake_policy()`` command to set it to ``OLD`` or ``NEW``
explicitly.

We use NO_POLICY_SCOPE to make the LXQtCompilerSettings commands
affect the includer.